pyproject.toml
[tool.poetry]
name = "fastapi-mongo-example"
version = "0.1.0"
description = "A simple app that uses FastAPI and MongoDB with some design patterns (such as dependency injection) to provide a simple example of CRUD application with that stack."
authors = ["greedWizard"]
license = "MIT"
readme = "README.md"
packages = [{include = "fastapi_mongo_example"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["uvicorn"], version = "^0.96.0"}
uvicorn = "^0.22.0"
ipython = "^8.14.0"
pymongo = "^4.3.3"
django-environ = "^0.10.0"
dependency-injector = "^4.41.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
